% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_downloads_stats.R
\name{get_download_stats}
\alias{get_download_stats}
\title{Get Neuroconductor download stats}
\usage{
get_download_stats(package = NULL, verbose = FALSE, error = FALSE)
}
\arguments{
\item{package}{Retrieve download stats for this package, if null return stats for all packages}

\item{verbose}{Get a compact or a detailed listing of downloads}

\item{error}{Should function error if httr::GET failed}
}
\value{
DataFrame with the download stats
}
\description{
Retrieve a compact or verbose list of downloads for one or all Neuroconductor packages
}
