% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_description.R
\name{get_package_description}
\alias{get_package_description}
\title{Retrieve Neuroconductor package DESCRIPTION file}
\usage{
get_package_description(package, error = FALSE)
}
\arguments{
\item{package}{Name of the package}

\item{error}{Should function error if httr::GET failed}
}
\value{
content of the DESCRIPTION file
}
\description{
Retrieves the DESCRIPTION file for a Neuroconductor package
}
