% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_details.R
\name{get_release_details}
\alias{get_release_details}
\title{Retrieve release details}
\usage{
get_release_details()
}
\value{
Content of the PACKAGES file associated with a specific release
}
\description{
Get full list of packages for a specific release
}
