% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_releases.R
\name{list_releases}
\alias{list_releases}
\title{List Neurocnductor releases}
\usage{
list_releases()
}
\value{
Data.frame of all releases
}
\description{
Retrieves the full list of Neuroconductor releases
}
