% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_list.R
\name{get_package_list}
\alias{get_package_list}
\title{Retrieve Neuroconductor package list}
\usage{
get_package_list()
}
\value{
DataFrame of Neuroconductor packages
}
\description{
Retrieves the complete list of Neuroconductor packages
}
